Fix using values() as defaults
I am using filter_data_server on a shiny app with several data frames.

I use:

```
mod <- datamods::filter_data_server(...)
mod$values()
```

To get the state of the filters. Afterwards, when the user requests the same dataset to be filtered I pass those `mod$values()` to the `defaults=` argument.

Unfortunately this approach fails when there is a sliderInput, I get the warning:

Warning: Error in : In sliderInput(): `min`, `max`, and `value` cannot be NULL, NA, or empty.